Refactor input handling: Remove mouse, keyboard, and touch modules
- Deleted the input.rs, keyboard.rs, mouse.rs, and touch.rs files to streamline the input handling system.
- Updated the WindowEvent enum to include documentation for the MovedOrResized variant.
- Enhanced layout module with additional documentation for CalculatedBounds and Constraints structs.
- Improved Position struct with detailed comments on its fields and methods.
- Added documentation for SvgElementType conversion method.

+++ engage-ux-core/src/modules/animation/animation_types.rs
@@ -109,8 +109,8 @@ impl AnimationType {
 			let p = p.max(0.0).min(1.0);
 			let lerp_f = |a: f32, b: f32| -> f32 { a + (b - a) * p };
 			Point {
-				x: lerp_f(ax, bx).round() as i64,
-				y: lerp_f(ay, by).round() as i64,
+				x: lerp_f(ax, bx).round(),
+				y: lerp_f(ay, by).round(),
 			}
 		};
